Add search and pagination
This commit adds filtering and pagination to the client library. All methods (except get_register_definition and get_custodian) now return a JSON response containing the particular results of the method that was called, in addition to the current page, number of total pages, number of results and a flag more_results to indicate if there are any more results to retrieve.

A new page_size option is added to the config_options object and defaults to 100 records.

Finally, the version is updated to 0.2.0.