Improve performance by restructuring downloaded data
This commit creates new `Item`, `Entry` and `Record` classes which each know how to parse a string of RSF which represents them. In combination with the changes made to the download of an RSF file when first creating a new `RegisterClient`, objects of the above classes only need to parse data when they are being accessed by the client, removing the burden of additional memory overhead until necessary.

Through this refactor, we improve the memory performance of creating a new `RegisterClient` quite substantially (see details on the related Trello card for details).