Download updates from registers
This commit introduces the ability for the client library to download only the updates from a register it has previously downloaded data for, rather than needing to download the entire register and re-parse it all again. To achieve this, we are using the latest `user` entry number currently read up-to by the last download, then downloading RSF from the register from this point.

The register data stored by the `RegisterClient` has been moved to an instance variable, due to a limitation of the MiniCache gem we are using which will wipe the data it has for a key if the cache expiration is deemed to have expired. Instead, we continue to use the MiniCache gem, but we use it as a means of controlling access to the underlying `@data` instance variable, so that it can continue to be updated when the cache expires, but without the side effects of wiping all data currently held and storing register data twice.