Get entries and record history allows entry number to be specified
This commit builds on the current `get_entries`, `get_records_with_history` and `get_metadata_records_with_history` methods, to allow a client to pass in an entry number if they would like to retrieve either entries or records that have occurred after this entry number. Through this extension, it is possible for clients to keep their own dataset up-to-date without needing to diff their data to the full data returned by the client library, or indeed to re-populate their entire dataset using the data from the client library.