Split out interface for data store
This commit refactors `RegisterClient` to split out the in-memory data store. This makes it possible for a client to pass in a custom data store implementing `DataStore` and only store the downloaded register data once, in a place convenient for the client. For example, the client may create an implementation of `DataStore` to interface with a local Postgres database, enabling the `RegisterClient` to act as an intermediary between registers and the downloaded register data.