Validate register proof and update tests
This commit introduces validation to check that any new RSF updates are valid for the current data. In practice this means that if the client library has read up to a certain point in the register and discovers that new updates are available, it will check to ensure that the total number of entries contained in the register is not less than what has currently been read up to, in addition to checking the `assert-root-hash` at the end of the incoming RSF to ensure this matches the current register proof of the register. If any of these checks fail, a new `InvalidRegisterError` is raised, with the message giving the details of the particular violation.