Correctly detect a reloaded remote register
The existing code does not correctly detect when a remote register
is inconsistent with the local register. This can happen in the
alpha and discovery environments where we reload registers and do
not follow the append-only rule.

Previously the code was comparing the last root hash in the RSF against
the current root hash of the remote register (this will almost always
match). Instead, the code should be comparing the first root hash of
the RSF against the current root hash of the local register. This is
the definition of that `assert-root-hash` command.

For now, we do not want to do the work in the library to compute
the root hash. As a temporary measure we will therefore set the
root hash value in the client by callign the `set_root_hash` method
in the DataStore. The proper thing to do is to remove this method and
calculate the root hash from the entries in the register, but we
can do this work later.

The comparison between latest entry number was removed, because if
the size of the register differs, the root-hashes will also differ.