Remove cache_duration property
This commit removes the `cache_duration` property from the `RegisterClientManager`. We no longer use `cache_duration` anywhere, as we instead use the implementation of `DataStore` to decide how and when data is updated. Specifically, our `InMemoryDataStore` caches data until the user calls `refresh_data` on the `RegisterClient` instance.

This commit also bumps the version of the client library to version `1.0.0`. This has been a long time in coming and will make it easier for us to properly manage versioning going forwards.