Ensure that system entries are not duplicated on refresh_data
/download-rsf/0 on an empty register always returns system entries.
For an empty register, the client library will containuously call
/download-rsf/0 because the parameter is based on user entry number
only. This means that currently, an empty register will have duplicated
system entries in the datastore if refresh_data is called on it multiple
times.

This fixes this issue by making sure that system entries are ignored
if the datastore contains at least one system entry.

This will not work for registers where system entries are added at
different times. However, this is a problem that already exists,
because ORJ only returns syste entries if you call /download-rsf/0 pr
download rsf for the whole register.

We can remove this "fix" when we merge the system and user entry logs.